Moved scores from flu build to base
References to builds directory from the base could will break with modular augur and they current break unit tests.
